Create InputData helper fn to import file or stdin
